fix: It finally works as expected
