--- packages/steam/AuthSchema.cs
@@ -1,9 +1,0 @@
-public class SteamUserCredentials
-{
-    public int Id { get; set; }  // Keep as primary key
-    public required string UserId { get; set; }
-    public required string AccountName { get; set; }
-    public required string RefreshToken { get; set; }
-    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-    public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;   
-}
