+++ packages/steam/SteamService.cs
@@ -1,201 +1,123 @@
-using SteamAuth;
 using Microsoft.EntityFrameworkCore;
 using System.Collections.Concurrent;
 
-// Steam Service
-public class SteamService
+namespace Steam
 {
-    private readonly ConcurrentDictionary<string, SteamAuthComponent> _clientHandlers = new();
-    private readonly IServiceProvider _serviceProvider;
-
-    public SteamService(IServiceProvider serviceProvider)
+    public class SteamService
     {
-        _serviceProvider = serviceProvider;
-    }
+        private readonly ConcurrentDictionary<string, SteamAuthComponent> _clientHandlers = new();
+        private readonly IServiceProvider _serviceProvider;
 
-    // In SteamService.cs
-    public Action SubscribeToEvents(string clientId, Action<ServerSentEvent> callback)
-    {
-        if (_clientHandlers.TryGetValue(clientId, out var handler))
+        public SteamService(IServiceProvider serviceProvider)
         {
-            var result = handler.Subscribe(callback);
-            return result;
+            _serviceProvider = serviceProvider;
         }
 
-        Console.WriteLine($"Warning: No handler found for client {clientId}");
-        return () => { }; // Empty unsubscribe function
-    }
-
-    public Action Subscribe(string userId, Action<string> callback)
-    {
-        if (_clientHandlers.TryGetValue(userId, out var handler))
+        public Action SubscribeToEvents(string userId, Action<ServerSentEvent> callback)
         {
-            return handler.Subscribe(callback);
+            if (_clientHandlers.TryGetValue(userId, out var handler))
+            {
+                return handler.Subscribe(callback);
+            }
+
+            Console.WriteLine($"Warning: No handler found for user {userId}");
+            return () => { }; // Empty unsubscribe function
         }
 
-        return () => { }; // Empty unsubscribe function
-    }
-
-    public async Task StartAuthentication(string userId)
-    {
-        // Check if we already have stored credentials
-        using var scope = _serviceProvider.CreateScope();
-        var dbContext = scope.ServiceProvider.GetRequiredService<SteamDbContext>();
-        var storedCredential = await dbContext.SteamUserCredentials
-            .FirstOrDefaultAsync(c => c.UserId == userId);
-
-        // Pass service provider instead of dbContext
-        var handler = _clientHandlers.GetOrAdd(userId, id => new SteamAuthComponent(_serviceProvider, userId,
-                    async (accountName, refreshToken) => await SaveCredentials(userId, accountName, refreshToken)));
-
-        if (storedCredential != null)
+        public async Task StartAuthentication(string userId)
         {
-            handler.SetCredentials(storedCredential.AccountName, storedCredential.RefreshToken);
-        }
-
-        await handler.HandleLoginRequest();
-        // handler.Disconnect();
-    }
-
-    public void DisconnectUser(string userId)
-    {
-        if (_clientHandlers.TryGetValue(userId, out var handler))
-        {
-            handler.Disconnect();
-            _clientHandlers.TryRemove(userId, out _);
-        }
-    }
-
-    private async Task SaveCredentials(string userId, string accountName, string refreshToken)
-    {
-        try
-        {
+            // Check for stored credentials
             using var scope = _serviceProvider.CreateScope();
             var dbContext = scope.ServiceProvider.GetRequiredService<SteamDbContext>();
-
-            var existingCredential = await dbContext.SteamUserCredentials
+            var storedCredential = await dbContext.SteamUserCredentials
                 .FirstOrDefaultAsync(c => c.UserId == userId);
 
-            if (existingCredential != null)
+            // Create or get auth component
+            var handler = _clientHandlers.GetOrAdd(userId, id =>
+                new SteamAuthComponent(_serviceProvider, userId,
+                    async (accountName, refreshToken) => await SaveCredentials(userId, accountName, refreshToken)));
+
+            // Set credentials if available
+            if (storedCredential != null)
             {
-                // Update existing record
-                existingCredential.AccountName = accountName;
-                existingCredential.RefreshToken = refreshToken;
-                existingCredential.UpdatedAt = DateTime.UtcNow;
+                handler.SetCredentials(storedCredential.AccountName, storedCredential.RefreshToken);
             }
-            else
+
+            // Start authentication process
+            await handler.HandleLoginRequest();
+        }
+
+        public void DisconnectUser(string userId)
+        {
+            if (_clientHandlers.TryGetValue(userId, out var handler))
             {
-                // Create new record
-                dbContext.SteamUserCredentials.Add(new SteamUserCredentials
+                handler.Disconnect();
+                _clientHandlers.TryRemove(userId, out _);
+            }
+        }
+
+        public void Unsubscribe(string userId, Action unsubscribeAction)
+        {
+            unsubscribeAction();
+        }
+
+        public async Task<SteamAccountInfo?> GetCachedUserInfoAsync(string userId)
+        {
+            if (string.IsNullOrEmpty(userId))
+                return null;
+
+            try
+            {
+                using var scope = _serviceProvider.CreateScope();
+                var dbContext = scope.ServiceProvider.GetRequiredService<SteamDbContext>();
+                return await dbContext.SteamAccountInfo
+                    .AsNoTracking()
+                    .FirstOrDefaultAsync(info => info.UserId == userId);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error retrieving cached user info: {ex.Message}");
+                return null;
+            }
+        }
+
+        private async Task SaveCredentials(string userId, string accountName, string refreshToken)
+        {
+            try
+            {
+                using var scope = _serviceProvider.CreateScope();
+                var dbContext = scope.ServiceProvider.GetRequiredService<SteamDbContext>();
+
+                var existingCredential = await dbContext.SteamUserCredentials
+                    .FirstOrDefaultAsync(c => c.UserId == userId);
+
+                if (existingCredential != null)
                 {
-                    UserId = userId,
-                    AccountName = accountName,
-                    RefreshToken = refreshToken
-                });
+                    // Update existing record
+                    existingCredential.AccountName = accountName;
+                    existingCredential.RefreshToken = refreshToken;
+                    existingCredential.UpdatedAt = DateTime.UtcNow;
+                }
+                else
+                {
+                    // Create new record
+                    dbContext.SteamUserCredentials.Add(new SteamUserCredentials
+                    {
+                        UserId = userId,
+                        AccountName = accountName,
+                        RefreshToken = refreshToken,
+                        CreatedAt = DateTime.UtcNow,
+                        UpdatedAt = DateTime.UtcNow
+                    });
+                }
+
+                await dbContext.SaveChangesAsync();
+                Console.WriteLine($"Saved Steam credentials for {userId}");
             }
-
-            await dbContext.SaveChangesAsync();
-            Console.WriteLine($"Saved Steam credentials for {userId}");
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Error saving credentials: {ex.Message}");
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error saving credentials: {ex.Message}");
+            }
         }
     }
-
-    // public async Task<SteamUserInfo?> GetUserInfoFromStoredCredentials(string userId)
-    // {
-    //     try
-    //     {
-    //         // Check if we have an active session
-    //         if (_clientHandlers.TryGetValue(userId, out var activeHandler) && activeHandler.UserInfo != null)
-    //         {
-    //             return activeHandler.UserInfo;
-    //         }
-
-    //         // Try to get stored credentials
-    //         using var scope = _serviceProvider.CreateScope();
-    //         var dbContext = scope.ServiceProvider.GetRequiredService<SteamDbContext>();
-    //         var storedCredential = await dbContext.SteamUserCredentials
-    //             .FirstOrDefaultAsync(c => c.UserId == userId);
-
-    //         if (storedCredential == null)
-    //         {
-    //             return null;
-    //         }
-
-    //         // Create a new handler with service provider instead of dbContext
-    //         var handler = new SteamClientHandler(userId, _serviceProvider, userId);
-    //         var success = await handler.LoginWithStoredCredentialsAsync(
-    //             storedCredential.AccountName,
-    //             storedCredential.RefreshToken
-    //         );
-
-    //         if (success)
-    //         {
-    //             _clientHandlers.TryAdd(userId, handler);
-    //             return handler.UserInfo;
-    //         }
-    //         return null;
-    //     }
-    //     catch (Exception ex)
-    //     {
-    //         Console.WriteLine($"Error retrieving cached user info: {ex.Message}");
-    //         return null;
-    //     }
-    // }
-
-    public async Task<SteamAccountInfo?> GetCachedUserInfoAsync(string userId)
-    {
-        if (string.IsNullOrEmpty(userId))
-            return null;
-
-        try
-        {
-            using var scope = _serviceProvider.CreateScope();
-            var dbContext = scope.ServiceProvider.GetRequiredService<SteamDbContext>();
-            return await dbContext.SteamAccountInfo
-                .AsNoTracking()
-                .FirstOrDefaultAsync(info => info.UserId == userId);
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Error retrieving cached user info: {ex.Message}");
-            return null;
-        }
-    }
-
-    public async Task<bool> HasCachedUserInfoAsync(string userId)
-    {
-        if (string.IsNullOrEmpty(userId))
-            return false;
-
-        try
-        {
-            using var scope = _serviceProvider.CreateScope();
-            var dbContext = scope.ServiceProvider.GetRequiredService<SteamDbContext>();
-            return await dbContext.SteamAccountInfo
-                .AnyAsync(info => info.UserId == userId);
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($"Error checking for cached user info: {ex.Message}");
-            return false;
-        }
-    }
-
-    public void Unsubscribe(string clientId, Action unsubscribeAction)
-    {
-        unsubscribeAction();
-    }
-
-    // public SteamUserInfo? GetUserInfo(string clientId)
-    // {
-    //     if (_clientHandlers.TryGetValue(clientId, out var handler))
-    //     {
-    //         return handler.UserInfo;
-    //     }
-
-    //     return null;
-    // }
 }
